Retrieve metrics from all pages
