+++ src/api/time-series.js
@@ -10,46 +10,48 @@ const TIME_SERIES_QUERY = `
     $to: String
   ) {
     organisation {
-      timeSeries(site: $site, from: $from, to: $to, measurements: $measurements, pages: $pages, profiles: $profiles) {
-        times {
-          name
-          snapshot
-          timestamp
-        }
+      site(slug: $site) {
+        timeSeries(from: $from, to: $to, measurements: $measurements, pages: $pages, profiles: $profiles) {
+          times {
+            name
+            snapshot
+            timestamp
+          }
 
-        series {
-          name
-          profile
-          page
-          measurement
-          values
-        }
+          series {
+            name
+            profile
+            page
+            measurement
+            values
+          }
 
-        pages {
-          uuid
-          name
-          url
-          canonical
-        }
+          pages {
+            uuid
+            name
+            url
+            canonical
+          }
 
-        testProfiles {
-          uuid
-          name
-          jsIsDisabled
-          adBlockerIsEnabled
-          hasDeviceEmulation
-          hasBandwidthEmulation
-          isMobile
-        }
+          testProfiles {
+            uuid
+            name
+            jsIsDisabled
+            adBlockerIsEnabled
+            hasDeviceEmulation
+            hasBandwidthEmulation
+            isMobile
+          }
 
-        measurements {
-          name
-          label
-          formatter
-          docsPath
-        }
+          measurements {
+            name
+            label
+            formatter
+            docsPath
+          }
 
-        csv
+          csv
+        }
       }
     }
   }
@@ -72,7 +74,7 @@ const list = async ({
     from,
     to
   })
-  return response.organisation.timeSeries
+  return response.organisation.site.timeSeries
 }
 
 module.exports = {
