+++ src/utils/cli.js
@@ -21,6 +21,19 @@ const options = {
   },
   cursor: {
     describe: 'The cursor to fetch records after'
+  },
+  pages: {
+    describe:
+      'A list of page uuids to return metrics for. eg: --pages=4a82662c-67dc-40cd-a461-6afc904260f3 51b3cb71-e534-4d48-842a-d8c9da672f55 or use multiple --pages for each page'
+  },
+  profiles: {
+    describe:
+      'A list of profile uuids to return metrics for. eg: --profiles=4a82662c-67dc-40cd-a461-6afc904260f3 51b3cb71-e534-4d48-842a-d8c9da672f55 or use multiple --profiles for each profile'
+  },
+  metrics: {
+    type: 'array',
+    describe:
+      'A list of metrics to return. eg: --metrics=first-meaningful-paint first-interactive or use multiple --metrics flags for each metric'
   }
 }
